+++ Sources/App/base/loot/LootTableEntry.swift
@@ -9,7 +9,7 @@ import Foundation
 
 public protocol LootTableEntry {
     var item : any ItemStack { get }
-    var amount_min : UInt16 { get }
-    var amount_max : UInt16 { get }
+    var amount_min : UInt { get }
+    var amount_max : UInt { get }
     var chance : UInt8 { get }
 }
